+++ tests/structural_map_test.py
@@ -2,7 +2,7 @@
 import pytest
 
 from mets_builder.digital_object import DigitalObject
-from mets_builder.metadata import MetadataFormat, MetadataType, MetadataBase
+from mets_builder.metadata import MetadataFormat, MetadataType, Metadata
 from mets_builder.structural_map import StructuralMap, StructuralMapDiv
 
 
@@ -94,7 +94,7 @@ def test_add_metadata_to_div():
     div = StructuralMapDiv(div_type="test_type")
     assert div.metadata == set()
 
-    metadata = MetadataBase(
+    metadata = Metadata(
         metadata_type=MetadataType.DESCRIPTIVE,
         metadata_format=MetadataFormat.OTHER,
         other_format="PAS-special",